Partial-body dose estimation (Dolphin's method)

Method based on the paper by Dolphin, G. W. (1969). Biological Dosimetry with Particular Reference to Chromosome Aberration Analysis: A Review of Methods. International Atomic Energy Agency (IAEA) Retrieved from https://inis.iaea.org/search/search.aspx?orig_q=RN:45029080.

Usage

estimate_partial_body_dolphin(
  num_cases,
  case_data,
  fit_coeffs,
  fit_var_cov_mat,
  conf_int = 0.95,
  protracted_g_value = 1,
  genome_factor = 1,
  gamma,
  aberr_module = c("dicentrics", "translocations", "micronuclei")
)

Arguments

num_cases

number of cases.

case_data

Case data in data frame form.

fit_coeffs

Fitting coefficients matrix.

fit_var_cov_mat

Fitting variance-covariance matrix.

conf_int

Confidence interval, 95% by default.

protracted_g_value

Protracted \(G(x)\) value.

genome_factor

Genomic conversion factor used in translocations, else 1.

gamma

Survival coefficient of irradiated cells.

aberr_module

Aberration module.

Value

List containing estimated doses data frame, observed fraction of cells scored which were irradiated, estimated fraction of irradiated blood data frame, AIC, and conf_int_* used.